corrected comment on pr

Related Issue: SPBK-4003
