+++ ui/src/components/routes/hits/view/HitViewer.tsx
@@ -241,7 +241,6 @@ const HitViewer: FC = () => {
             <Tab label={t('hit.viewer.details')} value="details" onClick={() => setTab('details')} />
 
             {hit?.howler.dossier?.map((lead, index) => (
-              // TODO : AG : HERE RELY THE GOAL.
               <Tab
                 // eslint-disable-next-line react/no-array-index-key
                 key={'lead:' + index}
