+++ ui/src/locales/en/translation.json
@@ -1,9 +1,9 @@
 {
   "*": "All values",
-  "actions.running": "Action \"{{action}}\" is executing.",
-  "actions.succeeded": "Action \"{{action}}\" completed successfully.",
-  "actions.skipped": "Action \"{{action}}\" was skipped: {{messages}}",
   "actions.error": "Action \"{{action}}\" had error(s): {{messages}}",
+  "actions.running": "Action \"{{action}}\" is executing.",
+  "actions.skipped": "Action \"{{action}}\" was skipped: {{messages}}",
+  "actions.succeeded": "Action \"{{action}}\" completed successfully.",
   "add": "Add",
   "adminmenu": "Admin menu",
   "adminmenu.config": "Configuration",
@@ -315,6 +315,8 @@
   "modal.rule.title": "Create a New Rule",
   "no.data": "No Data",
   "none": "None",
+  "notebook.opensPanel": "Opens notebook panel",
+  "notebook.title": "Notebook",
   "on": "on",
   "open": "Open",
   "operations.add_label": "Add Label",
@@ -418,6 +420,9 @@
   "personalization.reset_text": "Reset to site defaults",
   "personalization.showbreadcrumbs": "Show Breadcrumbs",
   "personalization.sticky": "Sticky Topbar",
+  "pivot.noDescription": "No description",
+  "pivot.openDossier": "Open Dossier",
+  "pivot.url": "URL:",
   "query": "Query",
   "quicksearch.aria": "search",
   "quicksearch.placeholder": "Search ...",
