+++ ui/src/locales/fr/translation.json
@@ -2,10 +2,10 @@
   "*": "Toutes les valeurs",
   "Protected B": "Protégé B",
   "Unclassified//Official Use Only": "Non classé//Réservé à des fins officielles",
-  "actions.running": "Action \"{{action}}\" s'exécute.",
-  "actions.succeeded": "Action \"{{action}}\" achevé avec succès.",
-  "actions.skipped": "Action \"{{action}}\" a été ignorée: {{messages}}",
   "actions.error": "Action \"{{action}}\" a rencontré une ou des erreur(s): {{messages}}",
+  "actions.running": "Action \"{{action}}\" s'exécute.",
+  "actions.skipped": "Action \"{{action}}\" a été ignorée: {{messages}}",
+  "actions.succeeded": "Action \"{{action}}\" achevé avec succès.",
   "add": "Ajouter",
   "adminmenu": "Menu d'administration",
   "adminmenu.config": "Configuration",
@@ -319,6 +319,8 @@
   "modal.rule.title": "Créer une nouvelle règle",
   "no.data": "Aucune donnée",
   "none": "Rien",
+  "notebook.opensPanel": "Ouvre le panneau des notes",
+  "notebook.title": "Carnet",
   "on": "sur",
   "open": "Ouvert",
   "operations.add_label": "Ajouter un label",
@@ -423,6 +425,9 @@
   "personalization.reset_text": "Réinitialiser les paramètres",
   "personalization.showbreadcrumbs": "Afficher fils d'ariane",
   "personalization.sticky": "Barre supérieure collante",
+  "pivot.noDescription": "Aucune description",
+  "pivot.openDossier": "Ouvrir le dossier",
+  "pivot.url": "URL :",
   "query": "Requête",
   "query.invalid": "Requête invalide",
   "quicksearch.aria": "recherche",
